requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
polars>=0.20.0

# Visualization
plotly>=5.18.0
seaborn>=0.13.0
matplotlib>=3.8.0

# ML/DL
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
catboost>=1.2.0
shap>=0.44.0
imbalanced-learn>=0.12.0

# Time Series & Survival
lifelines>=0.28.0

# MLOps
mlflow>=2.10.0

# API & Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0

# Data Validation
pandera>=0.18.0
great-expectations>=0.18.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# LLM
groq>=0.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
joblib>=1.3.0
tqdm>=4.66.0
